debug certain crash

+++ src/move_gen/pieces/rook.rs
@@ -6,7 +6,7 @@ use crate::move_gen::pieces::common_moves;
 pub fn possible_r(b: &Board, white: bool) -> Vec<Move> {
     let index = if white { 1 } else { 0 };
     let own = if white { b.get_white_pieces() } else { b.get_black_pieces() };
-    let opp = if white { b.get_black_pieces() - b.pieces[K_INDEX as usize] } else { b.get_white_pieces() - b.pieces[(K_INDEX + 1) as usize] };
+    let opp = if white { b.get_black_pieces() - b.pieces[K_INDEX] } else { b.get_white_pieces() - b.pieces[K_INDEX + 1] };
     let mut list: Vec<Move> = Vec::new();
     let rooks = b.pieces[(R_INDEX + index) as usize];
